New cursor setup
You can now add your own cursors to the game.

+++ Client.cs
@@ -1,4 +1,6 @@
-echo("--| ...System_Revamp being delt with... |---");
+$System_RevampVersion = "0.27-1 ALPHA stable";
+
+echo("\n--| ...System_Revamp being delt with... |--");
 if(!isObject(NewCursorGui))
     exec("./NewCursorGui.gui");
 if(isObject(NetGraphGui)) {
@@ -26,47 +28,67 @@ if(!isObject(NewCursorButton)) {
         OptGraphicsPane.add(%button);
 }
 
-function ResetCursorDefaults() { // This is to forcefully make sure our cursor doesn't delete itself.
-    DefaultCursor.delete(); // If it exists, delete it, then reset it.
+function resetCursorDefaults() { // This is to forcefully make sure our cursor doesn't delete itself.
+    if(isObject(DefaultCursor))
+        DefaultCursor.delete();
     if($platform $= "macos") {
-        new GuiCursor(DefaultCursor ) {
+        new GuiCursor(DefaultCursor) {
             hotSpot = "4 4";
             renderOffset = "0 0";
-            bitmapName = "./cursor/cursor" @ $Pref::Revamp::Cursor;
+            bitmapName = "./cursor/" @ $Pref::Revamp::Cursor;
         };
     } else {
         new GuiCursor(DefaultCursor) {
             hotSpot = "1 1";
             renderOffset = "0 0";
-            bitmapName = "./cursor/cursor" @ $Pref::Revamp::Cursor;
+            bitmapName = "./cursor/" @ $Pref::Revamp::Cursor;
         };
     }
     Canvas.setCursor(DefaultCursor); // Set the cursor to the cursor you've set.
 }
 
-function newCursorGui::onWake(%this)
-{
-    //%centerX = getWord($pref::Video::resolution,1) / 2 - getWord(%this.extent,1) / 2;
-    %p = $Pref::Video::Resolution;
-    CursorGui.position = getRandom(-900 - %p, 900 + %p) SPC getRandom(-900 - %p, 900 + %p);
-    CursorGui.tangoMoveTo("center", "1000", "elastic");
+function GuiWindowCtrl::MoveToCenter(%this) {
+    %this.position = getRandom(-900 - (%p = $Pref::Video::Resolution), 900 + %p) SPC getRandom(-900 - %p, 900 + %p);
+    %this.tangoMoveTo("center", "1000", "elastic");
 }
 
-function newCursorGui::onSleep(%this)
-{
-    //%centerX = getWord($pref::Video::resolution,1) / 2 - getWord(%this.extent,1) / 2;
-    %p = $Pref::Video::Resolution;
-    CursorGui.position = getRandom(-900 - %p, 900 + %p) SPC getRandom(-900 - %p, 900 + %p);
+function newCursorGui::onWake(%this) {
+    %n = "./cursor/*.png";
+    %position = "1 1";
+    for(%file=findFirstFile(getWord(%n, %i));%file !$= "";%file=findNextFile(getWord(%n, %i))) {
+        %name = getSubStr(%file, strPos(%file, "_") + 1, strPos(%file, ".") - strPos(%file, "_") - 1);
+        %id = getSubStr(%file, strPos(%file, "_") - 1, 1);
+        %rad = new GuiRadioCtrl(cusorTemp);
+        CursorGuiScroll.add(%rad);
+        %rad.position = %position;
+        %rad.setName("cursor" @ %id);
+        %rad.command = "NewCursorGui::setCursor(\"Cursor" @ %id @ "_" @ %name @ "\");";
+        %rad.text = %name;
+        %position = "1" SPC getWord(%position, 1) + 20;
+        NewCursorGui.cursorIds++;
+    }
+    CursorGuiHelp.command = "messageBoxOK(" @
+        "\"System_Revamp\'s guide to cursors\"," @
+        "\"To add a cursor to this list, make a png file, add it to \'S_Revamp.zip/cursor/\', and name it as such: CursorID_YourCursorName.png\");";
+    CursorGui.MoveToCenter();
 }
 
-function NewCursorGui::setCursor(%this, %cursor)
-{
+function newCursorGui::onSleep(%this) {
+    for(%i=0;%i < NewCursorGui.cursorIds;%i++) {
+        if(isObject(%cursor = cursor@%i))
+            %cursor.delete();
+    }
+    NewCursorGui.cursorIds=0;
+    CursorGui.position = getRandom(-900 - (%p = $Pref::Video::Resolution), 900 + %p) SPC getRandom(-900 - %p, 900 + %p);
+}
+
+function NewCursorGui::setCursor(%cursor) {
     %f = new FileObject();
     %f.openForWrite("config/revamp.txt");
-    %f.writeLine("Cursor " @ %cursor);
+    %f.writeLine(%cursor);
     %f.close(); %f.delete();
     echo("Attempting to set cursor to" SPC $Pref::Revamp::Cursor = %cursor);
-    ResetCursorDefaults();
+    resetCursorDefaults();
 }
 
 package Revamp {
@@ -75,10 +97,12 @@ package Revamp {
         if(isFile("config/revamp.txt")) {
             %f = new FileObject();
             %f.openForRead("config/revamp.txt");
-            $Pref::Revamp::Cursor = getWord(%f.readLine(), 1);
+            $Pref::Revamp::Cursor = %f.readLine();
             %f.close(); %f.delete();
             ResetCursorDefaults();
         }
     }
 };
 activatePackage(Revamp);
+
+echo("--| ...System_Revamp finished loading successfuly! |--\n");
